Feat: Add support for top-layer mirror in decoder state initializers
Top layer states from the encoder json stack now test against decoder architecture as part of the architecture verification. This includes removing the default behavior of concatenating bidirectional state tensors in the decoder, because this violates the property of a mirror'd architecture. The final encoder layer states are copied 1:1 to all decoder layer states.